refactor: move CruzPanel and related classes to new package structure
