+++ src/main/java/tp01/gui/ej1/VentanaCruz.java
@@ -1,4 +1,4 @@
-package tp01.gui;
+package tp01.gui.ej1;
 
 import javax.swing.JFrame;
